requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-countries==5.3.3
django-rest-framework==0.1.0
djangorestframework==3.10.0
geoip2==2.9.0
idna==2.8
maxminddb==1.4.1
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
ua-parser==0.8.0
urllib3==1.25.3
user-agents==2.0
